move common logic into common file
