+++ internal/controller/common_test.go
@@ -3,21 +3,42 @@ package controller
 import (
 	"context"
 	"testing"
+	"time"
 
 	numaflowv1 "github.com/numaproj/numaflow/pkg/apis/numaflow/v1alpha1"
 	numaflowversioned "github.com/numaproj/numaflow/pkg/client/clientset/versioned"
+	"github.com/numaproj/numaplane/internal/common"
+	apiv1 "github.com/numaproj/numaplane/pkg/apis/numaplane/v1alpha1"
 	"github.com/stretchr/testify/assert"
 	appsv1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	k8sclientgo "k8s.io/client-go/kubernetes"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
+func createPipelineRolloutInK8S(ctx context.Context, t *testing.T, numaplaneClient client.Client, pipelineRollout *apiv1.PipelineRollout) {
+	err := numaplaneClient.Create(ctx, pipelineRollout)
+	assert.NoError(t, err)
+	err = numaplaneClient.Status().Update(ctx, pipelineRollout)
+	assert.NoError(t, err)
+}
+
+func createPipelineInK8S(ctx context.Context, t *testing.T, numaflowClientSet *numaflowversioned.Clientset, pipeline *numaflowv1.Pipeline) {
+	resultPipeline, err := numaflowClientSet.NumaflowV1alpha1().Pipelines(defaultNamespace).Create(ctx, pipeline, metav1.CreateOptions{})
+	assert.NoError(t, err)
+	resultPipeline.Status = pipeline.Status
+
+	// updating the Status subresource is a separate operation
+	_, err = numaflowClientSet.NumaflowV1alpha1().Pipelines(defaultNamespace).UpdateStatus(ctx, resultPipeline, metav1.UpdateOptions{})
+	assert.NoError(t, err)
+}
+
 func createISBSvcInK8S(ctx context.Context, t *testing.T, numaflowClientSet *numaflowversioned.Clientset, isbsvc *numaflowv1.InterStepBufferService) {
-	updatedISBSvc, err := numaflowClientSet.NumaflowV1alpha1().InterStepBufferServices(defaultNamespace).Create(ctx, isbsvc, metav1.CreateOptions{})
+	resultISBSvc, err := numaflowClientSet.NumaflowV1alpha1().InterStepBufferServices(defaultNamespace).Create(ctx, isbsvc, metav1.CreateOptions{})
 	assert.NoError(t, err)
 	// update Status subresource
-	updatedISBSvc.Status = isbsvc.Status
-	_, err = numaflowClientSet.NumaflowV1alpha1().InterStepBufferServices(defaultNamespace).UpdateStatus(ctx, updatedISBSvc, metav1.UpdateOptions{})
+	resultISBSvc.Status = isbsvc.Status
+	_, err = numaflowClientSet.NumaflowV1alpha1().InterStepBufferServices(defaultNamespace).UpdateStatus(ctx, resultISBSvc, metav1.UpdateOptions{})
 	assert.NoError(t, err)
 }
 
@@ -30,12 +51,31 @@ func createStatefulSetInK8S(ctx context.Context, t *testing.T, k8sClientSet *k8s
 	assert.NoError(t, err)
 }
 
-func createPipelineInK8S(ctx context.Context, t *testing.T, numaflowClientSet *numaflowversioned.Clientset, pipeline *numaflowv1.Pipeline) {
-	updatedPipeline, err := numaflowClientSet.NumaflowV1alpha1().Pipelines(defaultNamespace).Create(ctx, pipeline, metav1.CreateOptions{})
+func createDeploymentInK8S(ctx context.Context, t *testing.T, k8sClientSet *k8sclientgo.Clientset, deployment *appsv1.Deployment) {
+	resultDeployment, err := k8sClientSet.AppsV1().Deployments(defaultNamespace).Create(ctx, deployment, metav1.CreateOptions{})
 	assert.NoError(t, err)
-	updatedPipeline.Status = pipeline.Status
+	resultDeployment.Status = deployment.Status
+	_, err = k8sClientSet.AppsV1().Deployments(defaultNamespace).UpdateStatus(ctx, resultDeployment, metav1.UpdateOptions{})
+	assert.NoError(t, err)
+}
 
-	// updating the Status subresource is a separate operation
-	_, err = numaflowClientSet.NumaflowV1alpha1().Pipelines(defaultNamespace).UpdateStatus(ctx, updatedPipeline, metav1.UpdateOptions{})
-	assert.NoError(t, err)
+func createDefaultPipelineOfPhase(phase numaflowv1.PipelinePhase) *numaflowv1.Pipeline {
+	return &numaflowv1.Pipeline{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:              defaultPipelineName,
+			Namespace:         defaultNamespace,
+			UID:               "some-uid",
+			CreationTimestamp: metav1.NewTime(time.Now()),
+			Generation:        1,
+			Labels: map[string]string{
+				common.LabelKeyISBServiceNameForPipeline:  defaultISBSvcRolloutName,
+				common.LabelKeyPipelineRolloutForPipeline: defaultPipelineRolloutName},
+		},
+		Spec: numaflowv1.PipelineSpec{
+			InterStepBufferServiceName: defaultISBSvcRolloutName,
+		},
+		Status: numaflowv1.PipelineStatus{
+			Phase: phase,
+		},
+	}
 }
